script/requirements.txt
websocket-client
netifaces
requests